requirements.txt
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.4.0
seaborn>=0.11.0
pyyaml>=6.0
pandas>=1.3.0
pytest==7.4.4